[NFC] Move TypeOrExtensionDecl out of PrintOptions.h.
