+++ lib/AST/ASTPrinter.cpp
@@ -63,23 +63,6 @@ void PrintOptions::clearSynthesizedExtension() {
   TransformContext.reset();
 }
 
-TypeOrExtensionDecl::TypeOrExtensionDecl(NominalTypeDecl *D) : Decl(D) {}
-TypeOrExtensionDecl::TypeOrExtensionDecl(ExtensionDecl *D) : Decl(D) {}
-
-Decl *TypeOrExtensionDecl::getAsDecl() const {
-  if (auto NTD = Decl.dyn_cast<NominalTypeDecl *>())
-    return NTD;
-
-  return Decl.get<ExtensionDecl *>();
-}
-DeclContext *TypeOrExtensionDecl::getAsDeclContext() const {
-  return getAsDecl()->getInnermostDeclContext();
-}
-NominalTypeDecl *TypeOrExtensionDecl::getBaseNominal() const {
-  return getAsDeclContext()->getAsNominalTypeOrNominalTypeExtensionContext();
-}
-bool TypeOrExtensionDecl::isNull() const { return Decl.isNull(); }
-
 TypeTransformContext::TypeTransformContext(Type T)
     : BaseType(T.getPointer()) {
   assert(T->mayHaveMembers());
